+++ lib/router.js
@@ -1878,7 +1878,7 @@ router.get('/simonsfoundation/articles', require('./routes/simonsfoundation/arti
 router.get('/simonsfoundation/recommend', require('./routes/simonsfoundation/recommend'));
 
 // 王者荣耀
-router.get('/pvp/newsindex/:type', require('./routes/pvp/newsindex'));
+router.get('/tencent/pvp/newsindex/:type', require('./routes/tencent/pvp/newsindex'));
 
 // 《明日方舟》游戏
 router.get('/arknights/news', require('./routes/arknights/news'));
